Fixed CustomSuits not existing

+++ sdk/clearCustomSuits.ts
@@ -1,5 +1,3 @@
-// Piepieonline's standard SMF ModSDK deployment script v1 
-
 import fs from "fs"
 import path from "path"
 
@@ -9,9 +7,13 @@ export const analysis = (context, api) => deploySDK.analysis(context, api);
 
 export const beforeDeploy = (context, api) => {
   const path = context.config.retailPath + '/mods/CustomSuits/';
-  fs.readdirSync(path).forEach(file => {
-    fs.unlinkSync(path + file);
-  })
+
+  if(fs.existsSync(path))
+  {
+    fs.readdirSync(path).forEach(file => {
+      fs.unlinkSync(path + file);
+    });
+  }
 
   api.logger.info('Custom suits cleared');
